backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.20.1+cpu
Pillow==10.0.0
huggingface_hub==0.20.0
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.26.0